pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask>=3.0.3",
    "flask-sqlalchemy>=3.1.1",
    "psycopg2-binary>=2.9.10",
    "google-generativeai>=0.8.3",
    "flask-login>=0.6.3",
    "flask-mail>=0.10.0",
    "apscheduler>=3.10.4",
    "feedparser>=6.0.11",
]
